+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - (Feature) Reconciliation Loop Interval option
 - (Bugfix) Fix GZIP encoding in case of small responses
 - (Bugfix) Fix PVC Rotation Discovery
+- (Feature) Allow to pass EphemeralStorage Resource to the Pods
 
 ## [1.2.32](https://github.com/arangodb/kube-arangodb/tree/1.2.32) (2023-08-07)
 - (Feature) Backup lifetime - remove Backup once its lifetime has been reached